transactionencoder replacing onehottransactions
