+++ mlxtend/preprocessing/__init__.py
@@ -12,8 +12,9 @@ from .dense_transformer import DenseTransformer
 from .copy_transformer import CopyTransformer
 from .onehot import one_hot
 from .onehot import OnehotTransactions
+from .transactionencoder import TransactionEncoder
 
 
 __all__ = ["MeanCenterer", "shuffle_arrays_unison", "CopyTransformer",
            "minmax_scaling", "standardize", "DenseTransformer",
-           "one_hot", "OnehotTransactions"]
+           "one_hot", "OnehotTransactions", "TransactionEncoder"]
